# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/partner/portal/login": {
      "post": {
        "operationId": "PortalController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Partner portal login — sets the session cookie",
        "tags": [
          "partner-portal"
        ]
      }
    },
    "/partner/portal/logout": {
      "post": {
        "operationId": "PortalController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "partner-portal"
        ]
      }
    },
    "/partner/portal/me": {
      "get": {
        "operationId": "PortalController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Current partner user + partner profile",
        "tags": [
          "partner-portal"
        ]
      }
    },
    "/partner/portal/dashboard": {
      "get": {
        "operationId": "PortalController_dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Partner metrics/summary widgets (own data only)",
        "tags": [
          "partner-portal"
        ]
      }
    },
    "/partner/portal/orders": {
      "get": {
        "operationId": "PortalController_orders",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "example": 50,
              "type": "string"
            }
          },
          {
            "name": "tripStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "List own orders (paginated, filterable)",
        "tags": [
          "partner-portal"
        ]
      }
    },
    "/partner/portal/orders/export": {
      "get": {
        "operationId": "PortalController_export",
        "parameters": [
          {
            "name": "format",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "xlsx",
                "pdf"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Export own orders (?format=pdf|xlsx)",
        "tags": [
          "partner-portal"
        ]
      }
    },
    "/partner/portal/orders/{id}": {
      "get": {
        "operationId": "PortalController_order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "One own order detail (403 for another partner’s order)",
        "tags": [
          "partner-portal"
        ]
      }
    },
...
  },
  "info": {
    "title": "fleet-taxi-dashboard-api",
    "description": "Fleet/deposit-reconciliation dashboard API. All money values are integer rupiah serialized as JSON numbers. Every response uses the standard envelope: `{ success: true, data, meta? }` on success, `{ success: false, error: { code, message, details? } }` on error.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sid"
      },
      "partner-api-key": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Partner API key (external /partner/v1 only)"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@fleet-taxi.id"
          },
          "password": {
            "type": "string",
            "example": "********"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  }
}